refactor!: separate into multiple files
BREAK CHANGE: services.addService(factory: (c) => Foo(c(Bar))) will no longer be used. Instead, use services.addService(factory: (c) => Foo(c.require<Bar>())).
